load-content.sh more idempotent

+++ load-content.sh
@@ -15,9 +15,12 @@ REGION=us-west-2
 filename="${UCLDC_EXHIBITIONS_DATA##*/}"         # http://unix.stackexchange.com/a/64435/40198
 name=$(echo $filename | rev | cut -c 4- | rev )  # http://stackoverflow.com/a/5863742/1763984
 
-aws s3 cp $UCLDC_EXHIBITIONS_DATA .
-gunzip $filename
-python manage.py loaddata $name
+if [[ ! -e $name ]]  # have we seen this one before
+  then
+    aws s3 cp $UCLDC_EXHIBITIONS_DATA .
+    gunzip $filename
+    python manage.py loaddata $name
+fi
 
 # Copyright (c) 2016, Regents of the University of California
 #
